[REF] mail: plain split
