+++ addons/im_livechat/__manifest__.py
@@ -94,6 +94,7 @@ Help your customers with this chat, and analyse their feedback.
             'web/static/lib/odoo_ui_icons/style.css',
             'web/static/src/scss/ui.scss',
             ('remove', 'web/static/src/core/browser/title_service.js'),
+            'mail/static/src/model/**/*',
             'mail/static/src/core/common/**/*',
             'mail/static/src/discuss/core/common/*',
             'mail/static/src/discuss/call/common/**',
