+++ addons/mail/static/src/core/common/store_service.js
@@ -1,4 +1,4 @@
-import { BaseStore, makeStore, Record } from "@mail/core/common/record";
+import { Store as BaseStore, makeStore, Record } from "@mail/core/common/record";
 import { reactive } from "@odoo/owl";
 
 import { router } from "@web/core/browser/router";
