+++ addons/mail/static/src/model/store.js
@@ -1,0 +1,10 @@
+export class BaseStore extends Record {
+    storeReady = false;
+    /**
+     * @param {string} localId
+     * @returns {Record}
+     */
+    get(localId) {
+        return this.recordByLocalId.get(localId);
+    }
+}
